Add StepFlow for steps navigation
This class controls the sequencing of the registered steps with the
methods `next_step_path` and `previous_step_path`.

The method `StepFlow.matches?` will be used at the routing level to
restrict access only to registered steps.

The method `StepFlow.current_step` instanciates a requested step in
the step controller