Add GenerateUrns service
This service will create all the available urns per application route.
They will be randomized, unique and the next available urn to be taken
will retreived with the Urn.next method.

Once all the urn are exhausted this method will raise an error and the
format of the URN will need updating ie increase the size of the
suffix set.  Urn::MAX_SUFFIX