Fix security audit dependencies
The external links in erb templates have been updated to open in a new tab using the 'new_tab: true' attribute instead of 'target: "_blank"'. This change was required to support the latest view_component gem that was required by the govuk_components gem. The versions of dependencies 'puma' and 'govuk-components' in the Gemfile have been updated to the latest to address the security advisories.